[account] Split account address tests, add coverage
